testing user auth, looks good to go ready to move on

+++ client/pages/item/all.js
@@ -4,21 +4,36 @@ import { useEffect, useState } from 'react';
 import Nav from '../components/Nav';
 import { Button } from '@mui/material';
 export default function All() {
+  const [user, setUser] = useState({});
+  const getUser = async () => {
+    const res = await axios.get('http://localhost:5000');
+    console.log(res.data);
+    setUser(res.data.user);
+  };
+
+  useEffect(() => {
+    getUser();
+  }, []);
+  const [errMsg, settErrMsg] = useState('');
   const [items, setItems] = useState([]);
   const fetchItems = async () => {
     const res = await axios.get('http://localhost:5000/item/all');
     setItems(res.data);
-    console.log(res.data);
+    // console.log(res.data);
   };
 
   useEffect(() => {
     fetchItems();
   }, []);
   async function handleDeleteItem(id) {
-    const res = await axios.delete(`http://localhost:5000/item/delete/${id}`);
-    // fetchItems();
-    setTimeout(fetchItems, 0);
-    console.log(res);
+    if (user && user.status === 'admin') {
+      const res = await axios.delete(`http://localhost:5000/item/delete/${id}`);
+      // fetchItems();
+      setTimeout(fetchItems, 0);
+      console.log(res);
+    } else {
+      settErrMsg('insufficient rights, please see an admin');
+    }
   }
   return (
     <div>
@@ -28,29 +43,35 @@ export default function All() {
         <Link href="/item/new">
           <Button variant="outlined">Create a new Item</Button>
         </Link>
-        {items.map(item => {
-          const url = `http://localhost:3000/item/${item.category.name}`;
-          return (
-            <div className="m-8" key={item._id}>
-              <Link href={url}>
-                <h2 className="cursor-pointer">{item.category.name}</h2>
-              </Link>
+        <h1>{errMsg}</h1>
+        <section className="border-8 border-blue-400 flex gap-4">
+          {items.map(item => {
+            const url = `http://localhost:3000/item/${item.category.name}`;
+            return (
+              <div className="m-8" key={item._id}>
+                <Link href={url}>
+                  <h2 className="cursor-pointer">{item.category.name}</h2>
+                </Link>
 
-              <h2>{item.model}</h2>
-              <h2>{item.Total_Value}</h2>
-              <h3 className="text-bold text-2xl">
-                ${item.quantity * item.price}
-              </h3>
-              <Button
-                color="error"
-                variant="outlined"
-                onClick={() => handleDeleteItem(item._id)}
-              >
-                Delete
-              </Button>
-            </div>
-          );
-        })}
+                <h2>{item.model}</h2>
+                <h2>{item.Total_Value}</h2>
+                <h3 className="text-bold text-2xl">
+                  ${item.quantity * item.price}
+                </h3>
+
+                {user && (
+                  <Button
+                    color="error"
+                    variant="outlined"
+                    onClick={() => handleDeleteItem(item._id)}
+                  >
+                    Delete
+                  </Button>
+                )}
+              </div>
+            );
+          })}
+        </section>
       </main>
     </div>
   );
